+++ index.html
@@ -13,12 +13,12 @@
 <body>
    <canvas id="canvas"></canvas>
    <script src="entities/Tile.js"></script>
-   <script src="providers/Provider.js"></script>
+   <script src="providers/TileProvider.js"></script>
    <script src="Parameters.js"></script>
-   <script src="providers/GlacialTilesProvider.js"></script>
-   <script src="providers/TemperateTilesProvider.js"></script>
-   <script src="services/TilesService.js"></script>
-   <script src="utils/Utils.js"></script>
+   <script type="module" src="providers/GlacialTilesProvider.js"></script>
+   <script type="module" src="providers/TemperateTilesProvider.js"></script>
+   <script type="module" src="services/TilesService.js"></script>
+   <script type="module" src="utils/Utils.js"></script>
    <script src="Perlin.js"></script>
    <script type="module" src="Main.js"></script>
    <script type="module" src="../youssef/lib-module.js"></script>
