+++ services/TilesService.js
@@ -1,8 +1,11 @@
+import {TemperateTilesProvider} from "../providers/TemperateTilesProvider.js";
+import {GlacialTilesProvider} from "../providers/GlacialTilesProvider.js";
+
 let providers = new Map();
 
 providers.set(-1, new GlacialTilesProvider());
 providers.set(0, new TemperateTilesProvider());
 
-function getTile(temperature, noise){
+export function getTile(temperature, noise){
     return providers.get(temperature).getTile(noise);
 }
